requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
google-generativeai